+++ lib/questionnaires/model/narrative_aggregator.dart
@@ -89,7 +89,7 @@ class NarrativeAggregator extends Aggregator<Narrative> {
               div.write('<p>${answer.valueInteger!.value}</p>');
             } else if (answer.valueCoding != null) {
               div.write(
-                  '<p>${answer.valueCoding!.localizedDisplay(locale)}</p>');
+                  '<p>- ${answer.valueCoding!.localizedDisplay(locale)}</p>');
             } else if (answer.valueDateTime != null) {
               div.write('<p>${answer.valueDateTime!.format(locale)}</p>');
             } else if (answer.valueDate != null) {